% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{app_sys}
\alias{app_sys}
\title{App Sys}
\usage{
app_sys(...)
}
\arguments{
\item{...}{(Optional) Character Vector(s) specifying paths and file(s)
within the R package's installation folder. Defaults to nothing which returns
the root of the app package folder.}
}
\value{
Character vector of positive length, containing the file paths that
  matched `...`, or the empty string, `""`, if none matched.
}
\description{
Access files in the current app.
}
\seealso{
[system.file()]
}
