% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshinycloudrun-package.R
\docType{package}
\name{rshinycloudrun-package}
\alias{rshinycloudrun}
\alias{rshinycloudrun-package}
\title{rshinycloudrun: Shiny App on Google Cloud Run}
\description{
Deploy a Shiny app to Google Cloud Run.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@noclocks.dev}

}
\keyword{internal}
